# [libc++] Vectorize mismatch