@philnik777 2024-02-24
This uses extensions, but I don't expect this to land that soon anyways. Worst case I'll have to wait a few weeks until I can land it.

CC @DenisYaroshevskiy in case you are interested in helping review this.