@DenisYaroshevskiy 2024-02-24
At the moment it's not trivial to run benchmarks for this against what I have. Is there something standalone?