@philnik777 2024-02-26
> At the moment it's not trivial to run benchmarks for this against what I have. Is there something standalone?

Could you elaborate a bit? I'm not sure what you're trying to do.