@philnik777 2024-02-27
> > Could you elaborate a bit? I'm not sure what you're trying to do.
> 
> I was thinking to compare this implementation on the benchmarks I have. But when it's a part of a libc++ fork it's a bit too much work.

With a bit of playing preprocessor (and not disabling the old `operation_traits.h`) you should be able to get things working with a 17 release: https://godbolt.org/z/j16P5Kxeb. Just put it before any other includes and everything should work (I think).