@DenisYaroshevskiy 2024-03-14
>I think we should use the same underlying machinery for both.

How'd you go about that? You need some compiler extensions to do predicates