@ldionne 2024-03-14
> > I think we should use the same underlying machinery for both.
> 
> How'd you go about that? You need some compiler extensions to do predicates

At the very least, `std::mismatch(std::execution::unseq, ...)` should be able to use the same optimizations as the serial version, so there's something to be shared. Not everything may be shareable, but some of it must be.